Final Scripts/requirements.txt
en-core-web-sm==2.1.0
enum34==1.1.2
Flask==1.0.2
graphviz==0.10.1
html5lib==1.0.1
Jinja2==2.10.1
jsonschema==2.6.0
lockfile==0.12.2
numpy==1.16.2
pandas==0.24.2
pathlib==1.0.1
Pillow==3.1.2
python-apt==1.1.0b1+ubuntu0.16.4.3
python-dateutil==2.8.0
python-debian==0.1.27
requests==2.21.0
spacy==2.1.3
thinc==7.0.4
toolz==0.9.0
tqdm==4.31.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.15.2
xlrd==1.2.0